+++ server/handlers/custom-route-handler.js
@@ -66,6 +66,7 @@ exports.customRouteHandler = function customRouteHandler(
     logError,
     seo,
     domainSlug,
+    cdnProvider = null
   }
 ) {
   const url = urlLib.parse(req.url, true);
@@ -84,7 +85,7 @@ exports.customRouteHandler = function customRouteHandler(
             "Defaulting the status-code to 302 with destination-path as home-page"
           );
         }
-        addCacheHeadersToResult(res, page.cacheKeys(config["publisher-id"]));
+        addCacheHeadersToResult({ res: res, cacheKeys: page.cacheKeys(config["publisher-id"]), cdnProvider: cdnProvider, config: config });
 
         let destination = page["destination-path"] || "/";
 
@@ -97,7 +98,7 @@ exports.customRouteHandler = function customRouteHandler(
       }
 
       if (page.type === "static-page") {
-        addCacheHeadersToResult(res, page.cacheKeys(config["publisher-id"]));
+        addCacheHeadersToResult({ res: res, cacheKeys: page.cacheKeys(config["publisher-id"]), cdnProvider: cdnProvider, config: config });
 
         if (page.metadata.header || page.metadata.footer) {
           return loadData("custom-static-page", {}, config, client, {