+++ server/handlers/story-redirect.js
@@ -6,14 +6,16 @@ exports.redirectStory = function redirectStory(
   req,
   res,
   next,
-  { logError, config, client }
+  { logError, config, client, cdnProvider = null }
 ) {
   return Story.getStoryBySlug(client, req.params.storySlug)
     .then((story) => {
       if (story) {
-        addCacheHeadersToResult(res, [
-          storyToCacheKey(config["publisher-id"], story),
-        ]);
+        addCacheHeadersToResult({
+          res: res, cacheKeys: [
+            storyToCacheKey(config["publisher-id"], story),
+          ], cdnProvider: cdnProvider, config: config
+        });
         return res.redirect(301, `/${story.slug}`);
       } else {
         return next();