+++ server/routes.js
@@ -141,10 +141,11 @@ function withConfigPartial(
 ) {
   return function withConfig(f, staticParams) {
     return function (req, res, next) {
+      const domainSlug = getDomainSlug(publisherConfig, req.hostname);
       const client = getClient(req.hostname);
       return client
         .getConfig()
-        .then((config) => configWrapper(config))
+        .then((config) => configWrapper(config, domainSlug))
         .then((config) =>
           f(
             req,
@@ -153,7 +154,7 @@ function withConfigPartial(
             Object.assign({}, staticParams, {
               config,
               client,
-              domainSlug: getDomainSlug(publisherConfig, req.hostname),
+              domainSlug,
             })
           )
         )
@@ -261,9 +262,11 @@ function getWithConfig(app, route, handler, opts = {}) {
  * @param {Object} opts.mobileConfigFields List of fields that are needed in the config field of the *&#47;mobile-data.json* API. This is primarily used by the React Native starter kit. (default: {})
  * @param {boolean} opts.templateOptions If set to true, then *&#47;template-options.json* will return a list of available components so that components can be sorted in the CMS. This reads data from *config/template-options.yml*. See [Adding a homepage component](https://developers.quintype.com/malibu/tutorial/adding-a-homepage-component) for more details
  * @param {boolean|function} opts.lightPages If set to true, then all story pages will render amp pages.
- * @param {string} opts.cdnProvider The name of the cdn provider. Supported cdn providers are akamai, cloudflare. Default value is cloudflare.
+ * @param {string | function} opts.cdnProvider The name of the cdn provider. Supported cdn providers are akamai, cloudflare. Default value is cloudflare.
  * @param {function} opts.maxConfigVersion An async function which resolves to a integer version of the config. This defaults to config.theme-attributes.cache-burst
  * @param {Array<object>|function} opts.redirectUrls An array or async function which used to render the redirect url provided in the array of object - >ex- REDIRECT_URLS = [{sourceUrl: "/tag/:tagSlug",destinationUrl: "/topic/:tagSlug",statusCode: 301,}]
+ * @param {boolean|function} redirectToLowercaseSlugs If set or evaluates to true, then for every story-page request having capital latin letters in the slug, it responds with a 301 redirect to the lowercase slug URL. (default: true)
+ * @param {boolean|function} shouldEncodeAmpUri If set to true, then for every story-page request the slug will be encoded, in case of a vernacular slug this should be set to false. Receives path as param (default: true)
  */
 exports.isomorphicRoutes = function isomorphicRoutes(
   app,
@@ -302,6 +305,8 @@ exports.isomorphicRoutes = function isomorphicRoutes(
     renderServiceWorker = renderServiceWorkerFn,
     publisherConfig = require("./publisher-config"),
     redirectUrls = [],
+    redirectToLowercaseSlugs = false,
+    shouldEncodeAmpUri,
   }
 ) {
   const withConfig = withConfigPartial(
@@ -377,6 +382,7 @@ exports.isomorphicRoutes = function isomorphicRoutes(
       seo,
       appVersion,
       cdnProvider,
+      redirectToLowercaseSlugs,
     })
   );
 
@@ -407,6 +413,7 @@ exports.isomorphicRoutes = function isomorphicRoutes(
         mobileApiEnabled,
         mobileConfigFields,
         cdnProvider,
+        redirectToLowercaseSlugs,
       })
     );
   }
@@ -456,17 +463,28 @@ exports.isomorphicRoutes = function isomorphicRoutes(
       cdnProvider,
       lightPages,
       redirectUrls,
+      redirectToLowercaseSlugs,
+      shouldEncodeAmpUri,
     })
   );
 
   if (redirectRootLevelStories) {
-    app.get("/:storySlug", withConfig(redirectStory, { logError }));
+    app.get(
+      "/:storySlug",
+      withConfig(redirectStory, { logError, cdnProvider })
+    );
   }
 
   if (handleCustomRoute) {
     app.get(
       "/*",
-      withConfig(customRouteHandler, { loadData, renderLayout, logError, seo })
+      withConfig(customRouteHandler, {
+        loadData,
+        renderLayout,
+        logError,
+        seo,
+        cdnProvider,
+      })
     );
   }